Use a simpler folder layout
This commit gets rid of the components subfolder and moves everything up into
src as explicitly named files instead of a lot of index.js files.
