+++ frontend/src/__tests__/Layout.test.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { render, cleanup } from '@testing-library/react'
-import { Layout } from '..'
+import { Layout } from '../Layout'
 
 describe('<Layout />', () => {
   beforeEach(() => (global.scrollTo = jest.fn()))
