+++ frontend/src/__tests__/TopBanner.test.js
@@ -2,28 +2,26 @@ import React from 'react'
 import { i18n } from '@lingui/core'
 import { I18nProvider } from '@lingui/react'
 import { ThemeProvider, theme } from '@chakra-ui/core'
-import { Footer } from '../'
+import { TopBanner } from '../TopBanner'
 import { render, cleanup } from '@testing-library/react'
 
 i18n.load('en', { en: {} })
 i18n.activate('en')
 
-describe('<Footer />', () => {
-  beforeEach(() => (global.scrollTo = jest.fn()))
+describe('<TopBanner />', () => {
   afterEach(cleanup)
 
-  it('renders children correctly', () => {
-    const { getAllByText } = render(
+  it('renders using the language prop correctly', () => {
+    const { getByRole } = render(
       <ThemeProvider theme={theme}>
         <I18nProvider i18n={i18n}>
-          <Footer>
-            <div>foo</div>
-          </Footer>
+          <TopBanner lang="en" />
         </I18nProvider>
       </ThemeProvider>,
     )
-
-    const test = getAllByText(/foo/)
-    expect(test).toHaveLength(1)
+    const test = getByRole('img')
+    expect(test.getAttribute('alt')).toBe(
+      'Symbol of the Government of Canada - Symbole du Gouvernement du Canada',
+    )
   })
 })
